requirements.txt
python-telegram-bot==20.0       # Основная библиотека для бота
requests==2.32.0                # Для отправки HTTP-запросов
schedule==1.1.0                 # Для периодического выполнения задач (например, обновление данных)
pandas==1.3.3                   # Для работы с данными, если понадобится
python-dotenv==0.19.0           # Для работы с переменными окружения

